:recycle: Refactor serialization of ShadowStyle
