+++ render-wasm/src/wasm.rs
@@ -3,6 +3,7 @@ pub mod fills;
 pub mod fonts;
 pub mod layouts;
 pub mod paths;
+pub mod shadows;
 pub mod shapes;
 pub mod strokes;
 pub mod text;
